# Switching deprecated buffer constructor to use Buffer.from