--- a/lib/encoding/bufferwriter.js
+++ b/lib/encoding/bufferwriter.js
@@ -4,9 +4,9 @@ var bufferUtil = require('../util/buffer');
 var assert = require('assert');
 
 var BufferWriter = function BufferWriter(obj) {
-  this.bufLen = 0;
   if (!(this instanceof BufferWriter))
     return new BufferWriter(obj);
+  this.bufLen = 0;
   if (obj)
     this.set(obj);
   else
@@ -15,7 +15,7 @@ var BufferWriter = function BufferWriter(obj) {
 
 BufferWriter.prototype.set = function(obj) {
   this.bufs = obj.bufs || this.bufs || [];
-  this.bufLen = this.bufs.reduce((prev, buf) => prev + buf.length, 0);
+  this.bufLen = this.bufs.reduce(function(prev, buf){ return prev + buf.length; }, 0);
   return this;
 };