--- a/lib/transaction/transaction.js
+++ b/lib/transaction/transaction.js
@@ -102,7 +102,11 @@ var hashProperty = {
   configurable: false,
   enumerable: true,
   get: function() {
-    return new BufferReader(this._getHash()).readReverse().toString('hex');
+    if(!this._hashCache) {
+      this._hash = new BufferReader(this._getHash()).readReverse().toString('hex');
+      this._hashCache = true;
+    }
+    return this._hash;
   }
 };
 Object.defineProperty(Transaction.prototype, 'hash', hashProperty);
@@ -838,6 +842,7 @@ Transaction.prototype._getInputAmount = function() {
 };
 
 Transaction.prototype._updateChangeOutput = function() {
+  this._hashCache = false;
   if (!this._changeScript) {
     return;
   }