--- a/lib/transaction/transaction.js
+++ b/lib/transaction/transaction.js
@@ -102,10 +102,7 @@ var hashProperty = {
   configurable: false,
   enumerable: true,
   get: function() {
-    if(!this._hashCache) {
-      this._hash = new BufferReader(this._getHash()).readReverse().toString('hex');
-      this._hashCache = true;
-    }
+    this._hash = new BufferReader(this._getHash()).readReverse().toString('hex');
     return this._hash;
   }
 };
@@ -842,7 +839,6 @@ Transaction.prototype._getInputAmount = function() {
 };
 
 Transaction.prototype._updateChangeOutput = function() {
-  this._hashCache = false;
   if (!this._changeScript) {
     return;
   }